Add generic pmove calculations for drawables
For anything that doesn't need special treatment (cgaz, snaphud),
'PmoveUtilsV2' now provides a generic pmove state that drawables can
use. The class is now a global object like the previous implementation,
and checks for any cvars that require the generic pmove to determine
whether the pmove is ran or not. It's still running a "local" copy
rather than using real pmove, as that is a requirement to eventually
implement 'g_synchronousClients'.
