+++ src/cgame/CMakeLists.txt
@@ -41,7 +41,7 @@ add_library(cgame MODULE
 	"cg_weapons.cpp"
 	"cg_window.cpp"
 	"etj_accel_color.cpp"
-	"etj_accel_color_data.cpp"
+	"etj_accel_color_v2.cpp"
 	"etj_accelmeter_drawable.cpp"
 	"etj_accelmeter_v2.cpp"
 	"etj_areaindicator_drawable.cpp"
@@ -101,7 +101,6 @@ add_library(cgame MODULE
 	"etj_speed_drawable.cpp"
 	"etj_spectatorinfo_data.cpp"
 	"etj_spectatorinfo_drawable.cpp"
-	"etj_strafe_quality_data.cpp"
 	"etj_strafe_quality_drawable.cpp"
 	"etj_strafe_quality_v2.cpp"
 	"etj_trace_utils.cpp"
