+++ src/cgame/etj_accelmeter_v2.cpp
@@ -34,10 +34,8 @@ inline constexpr float ACCEL_FOR_SOLID_COLOR = 100.0f;
 inline constexpr float ACCELMETER_SIZE_MIN = 1.0f;
 inline constexpr float ACCELMETER_SIZE_MAX = 10.0f;
 
-AccelMeterV2::AccelMeterV2(
-    const std::shared_ptr<AccelColorData> &accelColorData,
-    const std::shared_ptr<CvarUpdateHandler> &cvarUpdate)
-    : accelColorData(accelColorData), cvarUpdate(cvarUpdate) {
+AccelMeterV2::AccelMeterV2(const std::shared_ptr<CvarUpdateHandler> &cvarUpdate)
+    : cvarUpdate(cvarUpdate) {
   parseColor(etj_accelColor.string);
   setAccelColorStyle(etj_accelColorUsesAccel);
   setSize(etj_accelSize);
@@ -93,17 +91,17 @@ void AccelMeterV2::setAccelColorStyle(const vmCvar_t &cvar) {
   }
 }
 
-void AccelMeterV2::setupAccelColor(const AccelColorData::State &s,
-                                   const float speed, const vec2_t accel) {
+void AccelMeterV2::setupAccelColor(const PmoveUtilsV2::State &s,
+                                   const float speed, const vec2_t accelVec) {
   if (accelColorStyle == AccelColorStyle::NONE) {
     return;
   }
 
   if (accelColorStyle == AccelColorStyle::SIMPLE ||
-      accelColorData->lowSpeedOnGround(speed, s.pm.ps->groundEntityNum)) {
+      AccelColorV2::lowSpeedOnGround(speed, s.pm.ps->groundEntityNum)) {
     Vector4Copy(colorGreen, color);
 
-    float avgAccel = accelColorData->calcAvgAccel(storedSpeeds);
+    float avgAccel = AccelColorV2::calcAvgAccel(storedSpeeds);
 
     if (avgAccel < 0) {
       Vector4Copy(colorRed, color);
@@ -114,10 +112,10 @@ void AccelMeterV2::setupAccelColor(const AccelColorData::State &s,
 
     LerpColor(colorWhite, color, color, frac);
   } else if (accelColorStyle == AccelColorStyle::ADVANCED) {
-    if (accel[0] != 0 || accel[1] != 0) {
-      accelColorData->calcAdvancedAccelColor(s.pm, s.pml, accel, s.wishspeed,
-                                             s.wishvel, s.velAngle, s.optAngle,
-                                             color);
+    if (accelVec[0] != 0 || accelVec[1] != 0) {
+      AccelColorV2::calcAdvancedAccelColor(s.pm, s.pml, accelVec, s.wishspeed,
+                                           s.wishvel, s.velAngle, s.optAngle,
+                                           s.a, color);
     } else {
       Vector4Copy(colorWhite, color);
     }
@@ -139,7 +137,7 @@ bool AccelMeterV2::beforeRender() {
 
   lastUpdateTime = cg.snap->serverTime;
 
-  const AccelColorData::State &s = accelColorData->getState();
+  const PmoveUtilsV2::State &s = cgame.utils.pmoveV2->getState();
 
   // spectators cannot reliably measure acceleration
   // because velocity isn't snapped
@@ -150,19 +148,19 @@ bool AccelMeterV2::beforeRender() {
 
   if (accelColorStyle == AccelColorStyle::SIMPLE ||
       (accelColorStyle == AccelColorStyle::ADVANCED) &&
-          accelColorData->lowSpeedOnGround(s.vf, s.pm.ps->groundEntityNum)) {
+          AccelColorV2::lowSpeedOnGround(s.vf, s.pm.ps->groundEntityNum)) {
     storedSpeeds.push_back({lastUpdateTime, s.vf});
-    accelColorData->popOldStoredSpeeds(storedSpeeds, lastUpdateTime);
+    AccelColorV2::popOldStoredSpeeds(storedSpeeds, lastUpdateTime);
   } else if (!storedSpeeds.empty()) {
     storedSpeeds.clear();
   }
 
-  Vector2Subtract(s.pm.ps->velocity, lastSpeed, accel);
+  Vector2Subtract(s.pm.ps->velocity, lastSpeed, accelVec);
   Vector2Copy(s.pm.ps->velocity, lastSpeed);
 
   // FIXME: this calls color parsing every frame when noclip is active
   if (s.pm.ps->pm_type == PM_NOCLIP) {
-    Vector2Set(accel, 0, 0);
+    Vector2Set(accelVec, 0, 0);
 
     if (accelColorStyle != AccelColorStyle::NONE) {
       parseColor("white");
@@ -170,7 +168,7 @@ bool AccelMeterV2::beforeRender() {
       parseColor(etj_accelColor.string);
     }
   } else {
-    setupAccelColor(s, s.vf, accel);
+    setupAccelColor(s, s.vf, accelVec);
   }
 
   x = std::clamp(etj_accelX.value, 0.0f, 640.0f);
@@ -183,8 +181,8 @@ bool AccelMeterV2::beforeRender() {
   textStyle =
       etj_accelShadow.integer ? ITEM_TEXTSTYLE_SHADOWED : ITEM_TEXTSTYLE_NORMAL;
 
-  accelStr[0] = StringUtils::format("%.0f", accel[0]);
-  accelStr[1] = StringUtils::format("%.0f", accel[1]);
+  accelStr[0] = StringUtils::format("%.0f", accelVec[0]);
+  accelStr[1] = StringUtils::format("%.0f", accelVec[1]);
 
   return true;
 }
